+++ Sources/Sentry/SentryTransportAdapter.m
@@ -77,6 +77,21 @@ NS_ASSUME_NONNULL_BEGIN
     [self sendEnvelope:envelope];
 }
 
+- (void)storeEvent:(SentryEvent *)event traceContext:(nullable SentryTraceContext *)traceContext
+{
+    SentryEnvelopeItem *item = [[SentryEnvelopeItem alloc] initWithEvent:event];
+
+    SentryEnvelopeHeader *envelopeHeader = [[SentryEnvelopeHeader alloc] initWithId:event.eventId
+                                                                       traceContext:traceContext];
+
+    SentryEnvelope *envelope = [[SentryEnvelope alloc] initWithHeader:envelopeHeader
+                                                                items:@[ item ]];
+
+    for (id<SentryTransport> transport in self.transports) {
+        [transport storeEnvelope:envelope];
+    }
+}
+
 - (void)sendUserFeedback:(SentryUserFeedback *)userFeedback
 {
     SentryEnvelopeItem *item = [[SentryEnvelopeItem alloc] initWithUserFeedback:userFeedback];